pyproject.toml
[tool.poetry]
name = "graphene-jwt-sample"
version = "0.1.0"
description = ""
authors = ["madarame"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "3.2.0"
graphene = "^2.1.8"
graphene-django = "^2.15.0"
django-graphql-jwt = "0.3.1"
graphene-federation = "0.1.0"
django-extensions = "^3.1.3"
cryptography = "^3.4.7"
PyJWT = "2.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
